refactor : 기물의 이동과 관련된 테스트도 결국 Board의 메서드를 테스트하는 것이므로, board 테스트의 하위로 분류
