+++ src/main/java/domain/piece/bishop/Bishop.java
@@ -1,5 +1,6 @@
-package domain.piece;
+package domain.piece.bishop;
 
+import domain.piece.Piece;
 import domain.point.Direction;
 
 import java.util.HashMap;
