+++ src/main/java/domain/piece/knight/Knight.java
@@ -1,5 +1,6 @@
-package domain.piece;
+package domain.piece.knight;
 
+import domain.piece.Piece;
 import domain.point.Direction;
 
 import java.util.HashMap;
