+++ src/main/java/domain/piece/pawn/WhitePawn.java
@@ -1,0 +1,36 @@
+package domain.piece.pawn;
+
+import domain.piece.Piece;
+import domain.point.Direction;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class WhitePawn extends Piece {
+    @Override
+    public String getSymbol() {
+        return "p";
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        return getClass() == obj.getClass();
+    }
+
+    @Override
+    public Map<Direction, Integer> getMovableRange() {
+        Map<Direction, Integer> movableRange = new HashMap<>();
+        movableRange.put(Direction.DOWN, 2);
+        return movableRange;
+    }
+
+    @Override
+    public boolean isBlack() {
+        return false;
+    }
+
+    @Override
+    public boolean isWhite() {
+        return true;
+    }
+}
