+++ src/test/java/domain/board/KingTest.java
@@ -1,8 +1,12 @@
-package domain.piece;
+package domain.board;
 
 import domain.Board;
 import domain.exception.BlockedPathException;
 import domain.exception.InvalidDestinationPointException;
+import domain.piece.Empty;
+import domain.piece.Piece;
+import domain.piece.king.WhiteKing;
+import domain.piece.pawn.WhitePawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
